moved transfor asset to assetSelection from aquarius class
