+++ src/@utils/aquarius.ts
@@ -1,10 +1,5 @@
-import {
-  Asset,
-  LoggerInstance,
-  PublisherTrustedAlgorithm
-} from '@oceanprotocol/lib'
+import { Asset, LoggerInstance } from '@oceanprotocol/lib'
 import { AssetSelectionAsset } from '@shared/FormFields/AssetSelection'
-import { PriceList } from './subgraph'
 import axios, { CancelToken, AxiosResponse } from 'axios'
 import { OrdersData_orders as OrdersData } from '../@types/subgraph/OrdersData'
 import { metadataCacheUri } from '../../app.config'
@@ -12,9 +7,7 @@ import {
   SortDirectionOptions,
   SortTermOptions
 } from '../@types/aquarius/SearchQuery'
-import { getServiceByName } from './ddo'
-import { getAccessDetailsForAssets } from './accessDetailsAndPricing'
-import { AssetExtended } from 'src/@types/AssetExtended'
+import { transformAssetToAssetSelection } from './assetConvertor'
 
 export const MAXIMUM_NUMBER_OF_PAGES_WITH_RESULTS = 476
 
@@ -204,45 +197,6 @@ export async function retrieveDDOListByDIDs(
   }
 }
 
-export async function transformDDOToAssetSelection(
-  datasetProviderEndpoint: string,
-  assets: Asset[],
-  selectedAlgorithms?: PublisherTrustedAlgorithm[],
-  cancelToken?: CancelToken
-): Promise<AssetSelectionAsset[]> {
-  const extendedAssets: AssetExtended[] = await getAccessDetailsForAssets(
-    assets
-  )
-  const algorithmList: AssetSelectionAsset[] = []
-
-  for (const asset of extendedAssets) {
-    const algoComputeService = getServiceByName(asset, 'compute')
-
-    if (
-      asset?.accessDetails.price &&
-      algoComputeService?.serviceEndpoint === datasetProviderEndpoint
-    ) {
-      let selected = false
-      selectedAlgorithms?.forEach((algorithm: PublisherTrustedAlgorithm) => {
-        if (algorithm.did === asset.id) {
-          selected = true
-        }
-      })
-      const algorithmAsset: AssetSelectionAsset = {
-        did: asset.id,
-        name: asset.datatokens[0].name,
-        price: asset.accessDetails.price,
-        checked: selected,
-        symbol: asset.datatokens[0].symbol
-      }
-      selected
-        ? algorithmList.unshift(algorithmAsset)
-        : algorithmList.push(algorithmAsset)
-    }
-  }
-  return algorithmList
-}
-
 export async function getAlgorithmDatasetsForCompute(
   algorithmId: string,
   datasetProviderUri: string,
@@ -268,11 +222,10 @@ export async function getAlgorithmDatasetsForCompute(
 
   if (computeDatasets.totalResults === 0) return []
 
-  const datasets = await transformDDOToAssetSelection(
+  const datasets = await transformAssetToAssetSelection(
     datasetProviderUri,
     computeDatasets.results,
-    [],
-    cancelToken
+    []
   )
   return datasets
 }
